initialize: always run data migration substeps
If user runs through initialize and chooses to skip applying the data
migration scripts to uncover all pg_upgrade --check errors, and then
decides to re-run initialize and apply the scripts they will be skipped
unless we always run them.

The only downside to useing AlwaysRun with the conditional
nonInteractive inside vs. outside the substep is that when
non-interactive the substep shows up on the user interface as running
even though it is skipped. Putting the conditional outside the substep
is super clunky. What we really want is AlwaysRunConditionally, but that
doesn't really make sense logically.  For now this is okay since the
non-interactive is only used for testing purposes.

(cherry picked from commit 9232acb39ed0af7c2421ba200429541ad693e777)