Exclude pg_upgrade output folder when copying to primaries
After the coordinator is upgraded, it will be copied to the segments so
they can also upgrade. When this is done, we need to exclude the
pg_upgrade_output.d folder from being copied over. If this is not done,
segment upgrades will fail because when they will try to create this
folder that already exists from coordinator upgrade.

If the folder does not exist it should no-op. So this should still work
for 5 > 6 upgrade.

gpdb reference commit:
https://github.com/greenplum-db/gpdb/commit/ea9c5ff048495ad4d4f42a79f2f05635257ad23d