Add `gpupgrade check` to run checks relevant to logical upgrade.
As part of the process of supporting upgrading from GPDB6 to GPDB7 using
gpupgrade, we make a subset of pg_upgrade checks available to users to
look for GPDB6 objects that are incompatible with GPDB7. The failed
checks must be resolved before attempting a GPDB upgrade. Only the
checks that are relevant to upgrading using gpbackup are run.

pg_upgrade commit reference that implements this behavior:
https://github.com/greenplum-db/gpdb/pull/xxxxx