Use host flag when running pg_isolation2_regress
When installing from source, postgres expects Postgres sockets to be in
`/tmp`. isolation2's python database connector, psycopg2, defaults to
finding Postgres sockets to be in `/var/run/postgresql`. Set host to be
`/tmp` to make psycopg2 look in `/tmp` for postgres sockets. Otherwise
it may fail with the following error if PGHOST is not correctly set.

```
  File "/usr/lib64/python3.6/site-packages/psycopg2/__init__.py", line 130, in connect
    conn = _connect(dsn, connection_factory=connection_factory, **kwasync)
psycopg2.OperationalError: could not connect to server: No such file or directory
       Is the server running locally and accepting
       connections on Unix domain socket "/var/run/postgresql/.s.PGSQL.6000"?
```