Revert "Apply migration script statements in parallel"
With the changes to no longer drop partitioned indexes prior
to upgrade, the code to apply the index script statements in
parallel is no longer relevant. For other migration scripts
that may benefit from parallel execution, it would be better
served to first consider how to rework pg_dump/pg_upgrade/gpupgrade
to remove the need for the script altogether. If it's not possible
to do so, then consider reimplementing portions of this commit.

This reverts commit 9574615a0955b35f43db651ae1b9ef9ed9616af3.