Update --jobs flag to int32
Update the --jobs flag to int32 to remove some extraneous casts
throughout the code.

Remove the `jobs` parameter from the config in favor of passing it
as a parameter to each of the three steps. Users may want to specify a
different value for `initialize` vs. `execute` vs. `finalize`, and this
makes it possible and more intuitive.

Pass through the value to InitializeCreateClusterRequest, ExecuteRequest
and FinalizeRequest.

For InitializeCreateClusterRequest, it's a drop-in replacement for the
previous config file value.

For ExecuteRequest, we want to provide jobs during pg_upgrade of
the segments to be able to tune the number of parallel relfile
transfers. This feature is currently disabled, but we'll want this
knob to be available when we add it.

For FinalizeRequest, we want to execute reindex and rebuild
tsvector substeps before upgrading mirrors, and these substeps live in
commanders.Finalize and flow through the cli_to_hub gRPC path.
As a result, we add jobs to FinalizeRequest and use the request.GetJobs()
method to set jobs for ReindexInvalidIndexes and RebuildTSVectorTables.

uint isn't a valid scalar type for protobuf, so another reason to refactor
jobs to be an int32.

We also remove references for data migration apply and revert since the
requirement for parallel jobs is no longer necessary.