Rework how indexes are upgraded
pg_dump now properly dumps CREATE INDEX statements for child partition
indexes during binary upgrade. This allows for accurate recreation of
the index hierarchy of partition tables. See
greenplum-db/gpdb#17780

As a result, we no longer need to drop all of the indexes on partitioned
tables during initialize and recreate them during finalize. This saves a
significant amount of downtime for the user and reduces disruption of
the source cluster.

The new process is:

**Initialize:**

Perform pg_upgrade checks for bitmap, bpchar_pattern_ops, and AO
indexes. pg_upgrade will emit a WARNING and continue if indexes of these
types are found.

` Your installation contains <type> indexes. These indexes have
different internal formats between your old and new clusters so they
must be reindexed with the REINDEX command. After upgrading, you will be
given REINDEX instructions.  `

**Execute:**

Internally, pg_upgrade will upgrade the QD instance and invalidate these
indexes on the new QD as one of the final steps. It will emit a similar
WARNING as above and also provide the list of REINDEX commands as a SQL
script. The QEs (segments) upgrades are then by bootstrapped by copying
the new QD data directory, meaning the indexes are also invalidated on
the new segments.

During the relfilenode transfer, the files for these indexes are NOT
transferred because they need to be rebuilt anyway. This saves
significant time and bandwidth.

**Revert:**

Because no changes are made to the source cluster indexes, no action is
required to revert after either the initialize or execute steps.

**Finalize:**

Previously, bitmap and bpchar_pattern_ops indexes were marked as invalid
but there was no automated process in place to reindex them. In
addition, AO indexes weren't marked as invalid but were broken because
they aren't upgradeable due to a change in the on-disk aotid format. As
a result, trying to use the index on a debug build would hit a failed
assertion, as detailed by the existing tests.

Now, we include a finalize step to REINDEX these invalidated indexes in
parallel. Post finalize, all indexes in the new cluster are fully
functional.

Additionally, we handle a special case where a table with a tsvector
column has an index on that column. tsvector columns must be rebuilt in
the new cluster, see `old_8_3_rebuild_tsvector_tables` in pg_upgrade.
However, we weren't automating this step either and left it to the user
to run the SQL script output by pg_upgrade. This SQL would fail if the
tsvector column had an index because GPDB6 does not allow ALTER TYPE on
an indexed column. Now, we automate rebuilding these tsvector columns
with a new finalize substep `Substep_finalize_rebuild_tsvector`. If we
identify an index on the column, we first drop the index, then ALTER
TYPE, then recreate the index. This workflow is preferable to dropping
the index in the source cluster during initialize because we want to
minimize changes to the source cluster.

Data migration scripts:

With these changes, we can remove the `partitioned_tables_indexes`
migration scripts altogether.