Reorder REINDEXes so most recent partitions are done first
Reorder the REINDEXes so partition tables partitioned by date get their
most recently created partitions REINDEX'd first. This is a very
specific ask from a user to REINDEX "hot" tables first.

In actual practice this should not have a significant impact on overall
workflow unless a user is trying to run finalize and production at the
same time. At this point in upgrade finalize, the mirrors haven't been
created yet an the REINDEXes have not completed so we would not even
recommend that production workload be restarted.